requirements.txt
django==4.2
djangorestframework
beautifulsoup4
requests
redis
psycopg2-binary
django-redis